requirements.txt
requests==2.31.0
terminaltables==3.1.10
load-dotenv==0.1.0
python-dotenv==1.0.1